Shorted useless 0s in variables.

+++ 15in Worlds/include/atum8/globals.hpp
@@ -39,9 +39,9 @@ extern pros::Rotation leftEncoder;
 extern pros::Rotation backEncoder;
 
 //
-const double encoderWheelRadius{1.625}; // inches
+const float encoderWheelRadius{1.625}; // inches
 const double encoderWheelCircumference{200/25.4};
 
-const double driveGearRatio {60.00000000000000000 / 36.00000000000000000};
+const double driveGearRatio {60.0 / 36.0};
 extern int program;
 } // namespace atum8
